netdog: Write primary interface sysctls in wicked DHCP helper
This change removes the `prepare-primary-interface` service and moves
its functionality to the wicked helper `netdog install`.
