+++ packages/os/netdog-tmpfiles.conf
@@ -1,2 +1,3 @@
 d /var/lib/netdog 0700 root root -
+d /run/netdog 0700 root root -
 Z /var/lib/netdog 0700 root root -
