+++ packages/os/os.spec
@@ -43,8 +43,7 @@ Source115: link-kernel-modules.service.in
 Source116: load-kernel-modules.service.in
 Source117: cfsignal.service
 Source118: generate-network-config.service
-Source119: prepare-primary-interface.service
-Source120: reboot-if-required.service
+Source119: reboot-if-required.service
 
 # 2xx sources: tmpfilesd configs
 Source200: migration-tmpfiles.conf
@@ -403,7 +402,7 @@ install -d %{buildroot}%{_cross_unitdir}
 install -p -m 0644 \
   %{S:100} %{S:101} %{S:102} %{S:103} %{S:105} \
   %{S:106} %{S:107} %{S:110} %{S:111} %{S:112} \
-  %{S:113} %{S:114} %{S:118} %{S:119} %{S:120} \
+  %{S:113} %{S:114} %{S:118} %{S:119} \
   %{buildroot}%{_cross_unitdir}
 
 %if %{with nvidia_flavor}
@@ -453,7 +452,6 @@ install -p -m 0644 %{S:300} %{buildroot}%{_cross_udevrulesdir}/80-ephemeral-stor
 %{_cross_bindir}/netdog
 %{_cross_tmpfilesdir}/netdog.conf
 %{_cross_unitdir}/generate-network-config.service
-%{_cross_unitdir}/prepare-primary-interface.service
 
 %files -n %{_cross_os}corndog
 %{_cross_bindir}/corndog
