+++ sources/api/netdog/README.md
@@ -24,9 +24,6 @@ name of the interface, and valid options are "dhcp4" and "dhcp6".  A "?" may be 
 to signify that the lease for the protocol is optional and the system shouldn't wait for it.  A
 valid example: `netdog.default-interface=eno1:dhcp4,dhcp6?`.
 
-The subcommand `prepare-primary-interface` writes the default sysctls for the primary interface to
-file in `/etc/sysctl.d`, and then executes `systemd-sysctl` to apply them.
-
 The subcommand `write-resolv-conf` writes the resolv.conf, favoring DNS API settings and
 supplementing any missing settings with DNS settings from the primary interface's DHCP lease.  It
 is meant to be used as a restart command for DNS API settings.
