+++ sources/api/netdog/src/cli/mod.rs
@@ -2,7 +2,6 @@ pub(crate) mod generate_hostname;
 pub(crate) mod generate_net_config;
 pub(crate) mod install;
 pub(crate) mod node_ip;
-pub(crate) mod prepare_primary_interface;
 pub(crate) mod remove;
 pub(crate) mod set_hostname;
 pub(crate) mod write_resolv_conf;
@@ -11,7 +10,6 @@ pub(crate) use generate_hostname::GenerateHostnameArgs;
 pub(crate) use generate_net_config::GenerateNetConfigArgs;
 pub(crate) use install::InstallArgs;
 pub(crate) use node_ip::NodeIpArgs;
-pub(crate) use prepare_primary_interface::PreparePrimaryInterfaceArgs;
 pub(crate) use remove::RemoveArgs;
 use serde::{Deserialize, Serialize};
 pub(crate) use set_hostname::SetHostnameArgs;
